Generic networks selector
